+++ packages/convert/app/components/Probe.tsx
@@ -1,12 +1,5 @@
-import {
-	MediaParserAudioCodec,
-	MediaParserVideoCodec,
-	ParseMediaOnProgress,
-	TracksField,
-} from '@remotion/media-parser';
-import React, {useCallback, useMemo, useRef, useState} from 'react';
+import React, {useCallback, useMemo, useState} from 'react';
 import {Source} from '~/lib/convert-state';
-import {formatBytes} from '~/lib/format-bytes';
 import {useIsNarrow} from '~/lib/is-narrow';
 import {AudioTrackOverview} from './AudioTrackOverview';
 import {ContainerOverview} from './ContainerOverview';
@@ -19,54 +12,15 @@ import {Card, CardDescription, CardHeader, CardTitle} from './ui/card';
 import {ScrollArea} from './ui/scroll-area';
 import {Separator} from './ui/separator';
 import {Skeleton} from './ui/skeleton';
-import {useProbe} from './use-probe';
+import {ProbeResult} from './use-probe';
 
 export const Probe: React.FC<{
 	readonly src: Source;
 	readonly setProbeDetails: React.Dispatch<React.SetStateAction<boolean>>;
-	readonly setAudioCodec: React.Dispatch<
-		React.SetStateAction<MediaParserAudioCodec | null>
-	>;
-	readonly setVideoCodec: React.Dispatch<
-		React.SetStateAction<MediaParserVideoCodec | null>
-	>;
 	readonly probeDetails: boolean;
-	readonly onTracks: (tracks: TracksField) => void;
-	readonly onDuration: (duration: number | null) => void;
-}> = ({
-	src,
-	probeDetails,
-	setProbeDetails,
-	setAudioCodec,
-	setVideoCodec,
-	onTracks,
-	onDuration,
-}) => {
-	const videoThumbnailRef = useRef<VideoThumbnailRef>(null);
-
-	const onVideoThumbnail = useCallback((frame: VideoFrame) => {
-		videoThumbnailRef.current?.draw(frame);
-	}, []);
-
-	const onProgress: ParseMediaOnProgress = useCallback(
-		async ({bytes, percentage}) => {
-			await new Promise((resolve) => {
-				window.requestAnimationFrame(resolve);
-			});
-			const notDone = document.getElementById('not-done');
-			if (notDone) {
-				if (percentage === null) {
-					notDone.innerHTML = `${formatBytes(bytes)} read`;
-				} else {
-					notDone.innerHTML = `${Math.round(
-						percentage * 100,
-					)}% read (${formatBytes(bytes)})`;
-				}
-			}
-		},
-		[],
-	);
-
+	readonly probeResult: ProbeResult;
+	readonly videoThumbnailRef: React.RefObject<VideoThumbnailRef>;
+}> = ({src, probeDetails, setProbeDetails, probeResult, videoThumbnailRef}) => {
 	const {
 		audioCodec,
 		fps,
@@ -79,16 +33,7 @@ export const Probe: React.FC<{
 		durationInSeconds,
 		done,
 		error,
-	} = useProbe({
-		src,
-		onVideoThumbnail,
-		onAudioCodec: setAudioCodec,
-		onVideoCodec: setVideoCodec,
-		onTracks,
-		logLevel: 'verbose',
-		onProgress,
-		onDuration,
-	});
+	} = probeResult;
 
 	const onClick = useCallback(() => {
 		setProbeDetails((p) => !p);
